separate files and refactor code
